% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{SurvMM}
\alias{SurvMM}
\title{SurvMM: Survival Mixture Model}
\description{
This package provides a framework that can estimate mixture models based on nonparametric heteroscedastic models.
project work
}
